Separate NavItem from NavGroup
