+++ components/src/core/Drawer/index.ts
@@ -1,6 +1,7 @@
 export { DrawerComponent as Drawer, DrawerComponentProps } from './Drawer';
-export { DrawerHeader, DrawerHeaderProps } from './DrawerHeader';
-export { DrawerSubheader, DrawerSubheaderProps } from './DrawerSubheader';
-export { DrawerBody, DrawerBodyProps } from './DrawerBody';
-export { DrawerFooter, DrawerFooterProps } from './DrawerFooter';
-export { DrawerNavGroup, NavItem, DrawerNavGroupProps } from './DrawerNavGroup';
+export * from './DrawerHeader';
+export * from './DrawerSubheader';
+export * from './DrawerBody';
+export * from './DrawerFooter';
+export * from './DrawerNavGroup';
+export * from './DrawerNavItem';
